feat: Remove persona headings
The "Visitors" and "Maintainers" headings were my first attempt to clarify the difference in intended audiences for each part of the announcement.

But those headings only make sense if the reader thinks of himself/herself as a "visitor" or a ",maintainer".

Removing the "visitor" heading immediately places more focus on the content of the announcement since there's no longer an extraneous heading.

Replacing the "Maintainers" heading with "Add this badge to your project" both makes the section more actionable, and subtley opens the invitation to individuals who don't think of themselves as "maintainers".